Rename youtube images
It turns out that one of uBlock Origin's default included filter lists
has the following as a global filter item:
/images/youtube_

I was accidentally triggering this filter with my chosen filenames, so
the images weren't showing up for me on github. Now that I've renamed
them to /images/yt_, things should be fine.